homepage: consolidate structure of the different language versions
We have several versions (in several languages) of the homepage. They
are using a lot of duplicated code, in particular for the blog roll or
the video slider. This makes it difficult to keep the versions in sync:
there are already a few differences on the Chinese translation, because
e.g. the PR adding it was created before title anchors were added to the
English version.

Let's move most of the duplicated code to a common/Index module, from
where each version can pull it. Titles are passed as arguments where
necessary, so we can pass strings in the relevant language for the page.